Refactored a lot of tests.
