+++ package.json
@@ -7,7 +7,7 @@
     "dist": "babel -d lib/ src/",
     "watch": "babel --watch -d lib/ src/",
     "prepublish": "npm run dist",
-    "test": "nyc --reporter=lcov --require babel-register ./node_modules/.bin/_mocha"
+    "test": "nyc --reporter=lcov --require babel-register ./node_modules/.bin/_mocha --recursive test"
   },
   "repository": {
     "type": "git",
