+++ test/integration/fixtures/index.js
@@ -1,0 +1,8 @@
+/** @flow */
+
+export {
+  setupSchema
+} from './setupSchema'
+export {
+  createFixtureManager
+} from './createFixtureManager'
