--- test/schema.js
@@ -1,1 +1,0 @@
-import './schema/index'
