--- test/schema/index.js
@@ -1,1 +1,0 @@
-import './tablebuilder'
