--- test/smoke.js
@@ -1,1 +1,0 @@
-import './smoke/index'
