--- test/smoke/index.js
@@ -1,1 +1,0 @@
-import './schema'
